requirements.txt
beautifulsoup4==4.9.1
gtts==1.1.8
gtts-token==1.1.3
imutils==0.5.3
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pipenv==2020.6.2
pipwin==0.5.0
playsound==1.2.2
PyAudio @ file:///C:/Users/Naveen/pipwin/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycparser==2.20
pygame==1.9.6
PyGithub==1.51
pyjsparser==2.7.1
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
pyperclip==1.8.0
PyPrind==2.11.2
PyQt5==5.15.0
PyQt5-sip==12.8.0
pySmartDL==1.3.3
pytesseract==0.3.4
pyttsx3==2.87
pytube==9.6.0
pytube3==9.6.4
pytz==2020.1
pywin32==227
requests==2.23.0
requests-oauthlib==1.3.0
sounddevice==0.3.15
SpeechRecognition==3.8.1
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
wavio==0.0.4
wikipedia==1.4.0

aiml~=0.9.2
argparse~=1.4.0